requirements.txt
flask
flask-mysqldb
flask-login
pymysql
werkzeug
firebase-admin
python-dotenv